requirements.txt
absl-py==0.14.0
albumentations==1.0.3
antlr4-python3-runtime==4.8
appdirs==1.4.4
black==21.4b2
cachetools==4.2.3
carla==0.9.11
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.3
cloudpickle==2.0.0
configparser==5.1.0
cycler==0.10.0
Cython==0.29.24
dataclasses==0.6
decorator==5.1.0
detectron2==0.5+cu101
distro==1.5.0
docker-pycreds==0.4.0
easydict==1.9
elementpath==1.3.3
ephem==4.0.0.2
fire==0.4.0
future==0.18.2
fvcore==0.1.5.post20211015
gitdb==4.0.9
GitPython==3.1.24
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.41.0
h5py==3.4.0
hydra-core==1.1.1
idna==3.2
imageio==2.9.0
importlib-resources==5.2.2
iopath==0.1.8
kiwisolver==1.3.2
llvmlite==0.37.0
Markdown==3.3.4
matplotlib==3.4.3
mypy-extensions==0.4.3
networkx==2.2
numba==0.54.1
numpy==1.20.3
oauthlib==3.1.1
omegaconf==2.1.1
opencv-python==4.5.1.48
opencv-python-headless==4.5.3.56
packaging==20.9
pandas==1.3.3
pathspec==0.9.0
pathtools==0.1.2
Pillow==8.3.2
portalocker==2.3.2
promise==2.3
protobuf==3.18.0
psutil==5.8.0
py-trees==0.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pydot==1.4.2
pygame==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
regex==2021.10.8
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-build==0.11.1
scikit-image==0.18.3
scipy==1.7.1
seaborn==0.11.2
sentry-sdk==1.5.0
Shapely==1.6.4.post2
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
subprocess32==3.5.4
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
tifffile==2021.8.30
toml==0.10.2
torch==1.9.1
torchvision==0.10.1
tqdm==4.62.3
typing-extensions==3.7.4.3
urllib3==1.26.7
Wand==0.6.7
wandb==0.12.7
Werkzeug==2.0.1
xmlschema==1.0.18
yacs==0.1.8
yaspin==2.1.0
Yolov4-kl-demo==0.0.2
yolov5==5.0.10
zipp==3.6.0
